moving to curatedcontent package
