+++ src/test/java/com/wikia/webdriver/pageobjectsfactory/pageobject/mercury/curatedcontent/CuratedContentPageObject.java
@@ -1,7 +1,8 @@
-package com.wikia.webdriver.pageobjectsfactory.pageobject.mercury;
+package com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.curatedcontent;
 
 import com.wikia.webdriver.common.contentpatterns.MercuryMessages;
 import com.wikia.webdriver.common.logging.PageObjectLogging;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.BasePageObject;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -217,11 +218,11 @@ public class CuratedContentPageObject extends BasePageObject {
     List<String> itemsLabels = new ArrayList<>();
     boolean conflict = false;
 
-    for(WebElement element : curatedContentItems) {
+    for (WebElement element : curatedContentItems) {
       element = element.findElement(By.cssSelector("div.item-caption"));
       String label = element.getText();
 
-      if(itemsLabels.contains(label)) {
+      if (itemsLabels.contains(label)) {
         conflict = true;
         break;
       }
