+++ src/test/java/com/wikia/webdriver/testcases/mercurytests/curatedcontenttests/MainPageTests.java
@@ -6,8 +6,8 @@ import com.wikia.webdriver.common.contentpatterns.MercuryWikis;
 import com.wikia.webdriver.common.core.url.UrlChecker;
 import com.wikia.webdriver.common.logging.PageObjectLogging;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.CuratedContentPageObject;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.CuratedMainPagePageObject;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.curatedcontent.CuratedContentPageObject;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.curatedcontent.CuratedMainPagePageObject;
 
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
@@ -153,7 +153,7 @@ public class MainPageTests extends NewTestTemplate {
     //TODO: make function from this
     //Ticket: https://wikia-inc.atlassian.net/browse/CONCF-894
     for (PageElements element : PageElements.values()) {
-      if(PageElements.TRENDING_ARTICLES.name.equals(element.name)) {
+      if (PageElements.TRENDING_ARTICLES.name.equals(element.name)) {
         continue;
       }
